+++ playgrounds/vue/src/components/Slider.vue
@@ -10,25 +10,25 @@ interface SliderProps {
   max?: number;
   step?: number;
   required?: boolean;
-  input: string | number | undefined;
   errors: [string, ...string[]] | null;
   props: FieldElementProps;
 }
 
 const props = defineProps<SliderProps>();
+const model = defineModel<number | undefined>({ required: true });
 </script>
 
 <template>
   <div :class="['px-8 lg:px-10', props.class]">
     <InputLabel :name="props.props.name" :label="label" :required="required" />
     <input
+      v-model.number="model"
       v-bind="props.props"
       :id="props.props.name"
       class="w-full"
       :min="min"
       :max="max"
       :step="step"
-      :value="input"
       :required="required"
       :aria-invalid="!!errors"
       :aria-errormessage="`${props.props.name}-error`"