+++ packages/core/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Change implementation of `setFieldInput` to set `isTouched` to `true`
+
 ## v0.2.0 (July 27, 2025)
 
 - Fix bug in `resetItemState` by also resetting `.errors` signals